Added homework 4a
Test rewritten to use cropped_aster and cropped_sites. Cropped_sites is a new feature data-set which contains 17 points at various locations inside the extent of cropped_aster.

BUT there seems to be some error in 'Get Projection' although nothing was changed in the file. It uses Sites.shp data-set which is also unchanged.