[Breaking] Deprecate serializeJSON
It returns a Ptr{Void}, and feels sufficiently unsafe for people to work with. (I initially thought it might return a String.)

We don't have any tests for it, and I haven't seen widespread use for it through GDAL, so it's safer to deprecate this function until I've thought through how to make it accessible to users.