[Breaking] Return `missing` if the field is set but null. (#238)
* Return `missing` if the field is set but null.

Fixes #177 .

I think this aligns the behavior a lot better across the different file formats (see the test sets in test/test_tables.jl)

This behavior is implemented through `getfield(feature, i)` which makes use of `getfieldtype(feature, i)`. That way, we are aligned in behavior for field subtypes even for e.g. displaying the fields.

Because a lot hinges on distinguishing between whether a field is set versus whether it is null (see https://github.com/Toblerity/Fiona/issues/460#issue-240013079), I have also added support for `isfieldnull()`, `isfieldsetandnotnull()`, and `setfieldnull!()`.

* Add docstring references for GDAL/OGR RFCs

* Document the intended ArchGDAL behavior for NULL semantics

* Add tests and implementation for getfield with missing values

As mathieu has observed, `OGRUnsetMarker` and `OGRNullMarkerare` are mutually exclusive. We implement that case to return nothing if it ever comes up, but it is not possible for us to the corresponding code path.

* Update the implementation

* Implement setting fields to missing or nothing.

* Update the documentation to indicate the methods being discussed

* remove assertions

* Map unset fields to `nothing` and null values to `missing`

Because getdefault() is meant to be used only when setting fields for notnullable columns with missing values, we make it return `nothing` instead of `missing` to unset the field.

* Turn references into hyperlinks

* use isfieldset() and isfieldnull() when getting a field

isfieldsetandnotnull() is just a convenience function.

* raise an error instead of getting default if we do not recognize the fieldtype when reading the field

(in the future, we will switch to dispatch on the fieldtype rather than the current dictionary approach, but that is out of the scope of this commit.)

* Fix docstring for isfieldsetandnotnull()

* Update to reflect NA for reading unset & null fields